+++ worker/pkg/workflows/datasync/activities/shared/shared.go
@@ -11,7 +11,7 @@ import (
 	sqlmanager_shared "github.com/nucleuscloud/neosync/backend/pkg/sqlmanager/shared"
 	neosync_benthos "github.com/nucleuscloud/neosync/worker/pkg/benthos"
 
-	http_client "github.com/nucleuscloud/neosync/worker/pkg/http/client"
+	http_client "github.com/nucleuscloud/neosync/internal/http/client"
 	"github.com/spf13/viper"
 )
